Move the shiftplanning api client to a separate project.
Note: In order to run the API it is now required to create a local Nuget-Package feed.
When the folder for the feed is created, the ShiftPlanningApiConnection should be packed and packaged into the nuget-feed.
In Visual Studio, add the nuget-feed in Nuget Package Manager -> Package sources.
Then you can add the ApiConnector using nuget (remember to choose the correct package source in the package manager).